make space for external updates in zone serial
As we might need ACME updates to zone files that are done externally,
there needs to be allowance for interim serial increments. I solved
this by appending 00 to the generated serials and limiting the
year to two digits.

Serials are uint32. This scheme will therefore be sufficient for almost the
next 20 years... We will need to see after that.